// webapp/app/containers/Dashboard/types.ts
export interface IDashboardConfig {
  filters: IControl[]
  linkages: any[]
  queryMode: ControlQueryMode
}